Change time attributes to long precision

+++ src/main/java/com/jarslab/ts/DataPoint.java
@@ -2,7 +2,7 @@ package com.jarslab.ts;
 
 public interface DataPoint
 {
-    int getTime();
+    long getTime();
 
     double getValue();
 }
